Rename to context package and refactor to allow more context props
