+++ packages/context/README.md
@@ -1,10 +1,12 @@
-# Theme
+# Context
 
-Provides theming capabilities by utilising React's Context API.
+Provides context capabilities by utilising React's Context API.
 
-This package should be used to create themes to be shared between multiple
-components, such as app-wide font size changes that defines the scale of the
-text views.
+This package should be used to create app-wide shared components, such as themes
+or shared callbacks.
+
+(See https://reactjs.org/docs/context.html on when to use context and when to
+avoid)
 
 ## Contributing
 
@@ -44,3 +46,5 @@ yarn test:all
 
 More theme properties will be added in the future, such as the section colours,
 that will be defined by the apps' current navigation state.
+
+Client callbacks that are bridged through NativeModules can utilise contexts.
