+++ packages/context/src/defaults/index.js
@@ -1,5 +1,7 @@
 import scales from "./scales";
 
 export default {
-  scale: scales.medium
+  theme: {
+    scale: scales.medium
+  }
 };
